Update code formatting and fix bug in file processing
